docs: enable TOC in pandoc
Pandoc is able to generate a much nicer ToC for us.

However, Github does not know how to generate a table of contents, and
so currently we help Github by adding in the table of contents
manually ourselves.

As such, with pandoc's "toc" option enabled, we will get two table of
contents, which is not very good.

Thus, we use our "githubonly" filter to filter out the manual table
of contents so that only Github will display it.