docs: make appendices true LaTeX appendices
Our Appendix sections are currently just "sections", which is not very
good because in our PDF output they are numbered just like any other
section.

We are able to tell LaTeX that the following sections are appendices by
using the "\appendix" command. So, let's put that in.

We wrap the "\appendix" around a "<!-- BEGIN LATEX" "END LATEX-->" block
so that Github does not render it.

We also need to make all the appendix headers level 1 headers, because
LaTeX requires that all appendix sections start with a level 1 header.