ObservableListChangeListener: use atomic operations
There is a bug where even though the task book has changed, the task
book was not being saved at all.

Presently (in LogicManager.java), we use a TaskBookChangeListener (which
inherits from ObservableListChangeListener) to monitor a task book from
changes. However, seemingly despite the fact that the task book was
modified, the ObservableListChangeListener was still returning
hasChanged = false!

It turns out that change listeners on an observable list may be called
from a different thread, and this was causing memory synchronization
problems. Fix this by using atomic operations instead.